# Improve performance of getData retrieval