--- a/lib/internal/Magento/Framework/DataObject.php
+++ b/lib/internal/Magento/Framework/DataObject.php
@@ -128,11 +128,10 @@ class DataObject implements \ArrayAccess
             return $this->_data;
         }
 
-        /* process a/b/c key as ['a']['b']['c'] */
-        if ($key !== null && strpos($key, '/') !== false) {
+        $data = $this->_data[$key] ?? null;
+        if ($data === null && $key !== null && strpos($key, '/') !== false) {
+           /* process a/b/c key as ['a']['b']['c'] */
             $data = $this->getDataByPath($key);
-        } else {
-            $data = $this->_getData($key);
         }
 
         if ($index !== null) {