@igorwulff 2022-10-17
@magento run all tests