@igorwulff 2022-10-19
@ihor-sviziev thanks, I assume this will get tested further and merged if ok? 

I'm also looking into other segments of this DataObject class and see more room for improvement especially the magic __call class, which is called a lot and costs quite a bit of performance. But I think we shold move that to a different potential PR? I'm just curious how you think about these potential changes (since they are a bit radical):

```
Index: vendor/magento/framework/DataObject.php
IDEA additional info:
Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
<+>UTF-8
===================================================================
diff --git a/vendor/magento/framework/DataObject.php b/vendor/magento/framework/DataObject.php
--- a/vendor/magento/framework/DataObject.php	
+++ b/vendor/magento/framework/DataObject.php	(date 1666202240469)
@@ -57,7 +57,7 @@
         }
 
         foreach ($arr as $index => $value) {
-            $this->setData($index, $value);
+            $this->_data[$index] = $value;
         }
         return $this;
     }
@@ -386,22 +386,26 @@
      */
     public function __call($method, $args)
     {
-        switch (substr($method, 0, 3)) {
-            case 'get':
-                $key = $this->_underscore(substr($method, 3));
-                $index = isset($args[0]) ? $args[0] : null;
-                return $this->getData($key, $index);
-            case 'set':
-                $key = $this->_underscore(substr($method, 3));
-                $value = isset($args[0]) ? $args[0] : null;
-                return $this->setData($key, $value);
-            case 'uns':
-                $key = $this->_underscore(substr($method, 3));
-                return $this->unsetData($key);
-            case 'has':
-                $key = $this->_underscore(substr($method, 3));
-                return isset($this->_data[$key]);
+        switch($method[0]) {
+            case 'g':
+                return $this->_data[
+                    self::$_underscoreCache[$method] ?? $this->_underscore($method)
+                ] ?? null;
+            case 's':
+                $this->_data[
+                    self::$_underscoreCache[$method] ?? $this->_underscore($method)
+                ] = $args[0] ?? null;
+                return $this;
+            case 'u':
+                return $this->unsetData(self::$_underscoreCache[$method] ?? $this->_underscore($method));
+            case 'h':
+                return isset(
+                    $this->_data[
+                        self::$_underscoreCache[$method] ?? $this->_underscore($method)
+                    ]
+                );
         }
+
         throw new \Magento\Framework\Exception\LocalizedException(
             new \Magento\Framework\Phrase('Invalid method %1::%2', [get_class($this), $method])
         );
@@ -434,7 +438,16 @@
         if (isset(self::$_underscoreCache[$name])) {
             return self::$_underscoreCache[$name];
         }
-        $result = strtolower(trim(preg_replace('/([A-Z]|[0-9]+)/', "_$1", $name), '_'));
+        $result = strtolower(
+            trim(
+                preg_replace(
+                    '/([A-Z]|[0-9]+)/', "_$1",
+                    lcfirst(substr($name, 3))
+                ),
+                '_'
+            )
+        );
+
         self::$_underscoreCache[$name] = $result;
         return $result;
     }

```