@engcom-Echo 2023-02-09
@igorwulff @ihor-sviziev 
Thanks for the collaboration & contribution!


:heavy_check_mark: QA Passed
Improve performance of getData retrieval

**Manual Scenario Steps**
1. Set magento on production mode with block_html disabled and enabled.
2. check the load time of PLP with sample data/without sample data

**Before:** :heavy_multiplication_x: 
Verified with sample data on different PLP

with block_html enabled
```
Pants women - 1.05s        
Tees men - 326ms
Watches -  792ms
Tees-women - 4.54s     
```

with block_html disabled 
```
Pants women - 1.97s
Watches - 1.96s
Tees-women - 1.21s
```

**After:** :heavy_check_mark: 
Verified with sample data on different PLP
with block_html enabled
```
Pants women - 868ms       
Tees men - 126ms
Watches -  463ms
Tees-women - 3.25s     
```

with block_html disabled 
```
Pants women - 1.79s
Watches -  698ms
Tees-women - 1.04s   
```
Hence moving to `Merge In Progress`